refactor(echo-test): messages -> msg_options
