+++ jarust_plugins/src/echotest/mod.rs
@@ -1,4 +1,4 @@
 pub mod events;
 pub mod handle;
 pub mod jahandle_ext;
-pub mod messages;
+pub mod msg_options;
